lib: misrify
JIRA: DO-3

+++ lib/cbuffer.h
@@ -52,13 +52,13 @@ static inline size_t _cbuffer_discard(cbuffer_t *buf, size_t sz)
 void _cbuffer_init(cbuffer_t *buf, void *data, size_t sz);
 
 
-unsigned int _cbuffer_read(cbuffer_t *buf, void *data, size_t sz);
+size_t _cbuffer_read(cbuffer_t *buf, void *data, size_t sz);
 
 
-unsigned int _cbuffer_write(cbuffer_t *buf, const void *data, size_t sz);
+size_t _cbuffer_write(cbuffer_t *buf, const void *data, size_t sz);
 
 
-unsigned int _cbuffer_peek(const cbuffer_t *buf, void *data, size_t sz);
+size_t _cbuffer_peek(const cbuffer_t *buf, void *data, size_t sz);
 
 
 #endif
